backend/requirements.txt
# Bexy Flowers AI Backend - Python Dependencies
# =============================================
# Install with: pip install -r requirements.txt

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# AI/ML Libraries
torch==2.1.2
torchvision==0.16.2
diffusers==0.25.0
transformers==4.36.2
accelerate==0.25.0

# Image Processing
Pillow==10.2.0

# Utilities
numpy==1.26.3
requests==2.31.0

# Optional: For faster loading and better performance
# xformers==0.0.23  # Uncomment for faster attention (GPU only, requires compilation)
# bitsandbytes==0.41.3  # Uncomment for 8-bit optimization (reduces VRAM usage)

